+++ bciers/apps/registration/app/data/jsonSchema/operationRegistration/operationInformation.ts
